Rename loader -> scenario
