+++ malsim/scenario.py
